Move components to separate folder and fix namespaces
