+++ PokemonCalcMauiBlazor/Pages/Attackdex.razor
@@ -2,6 +2,7 @@
 @inherits BasePage
 @inject IDialogService DialogService
 @inject System.IO.Abstractions.IFileSystem FileSystemAbstraction
+@using PkmnCalcMauiBlazor.Pages.Components
 
 <DexSelect Label="Attackdex Type" DexTypes="attackdexTypes"
            @bind-SelectedDexType="SelectedAttackdexType" @bind-DexSource="AttackdexSource" />
