+++ PokemonCalcMauiBlazor/Pages/Components/PokemonTypeCalculator.razor
