+++ PokemonCalcMauiBlazor/Pages/Components/PokemonTypeSelect.razor
