+++ PokemonCalcMauiBlazor/Pages/Components/SavePokemonNamesProgress.razor
