+++ PokemonCalcMauiBlazor/Pages/Index.razor
@@ -1,5 +1,6 @@
 ﻿@*// Copyright (c) 2022 iyarashii @ https://github.com/iyarashii
 // Licensed under the GNU General Public License v3.0.*@
 @page "/"
+@using PkmnCalcMauiBlazor.Pages.Components
 
 <PokemonTypeCalculator />
